fix: resolve audit, Windows build, and clippy CI failures
Three CI jobs were failing on master; this addresses all of them.

Security Audit: rustsec/audit-check posts its results as a GitHub
check-run, which requires `checks: write`. The workflow only granted
`contents: read`, so the action failed with "Resource not accessible
by integration" even when no vulnerabilities were found. The two yanked
`spin` advisories remain as non-blocking annotations (informational,
not vulnerabilities).

Windows build: linking all ~102 AWS SDK crates under --all-features
exhausted the runner's default page file (os error 1455 -> E0786 on
rlib mmap). Added a Windows-only page-file step before the Rust setup.

Lint/Clippy: the new `-D warnings` gate surfaced 35 lints never caught
before. Fixed mechanically — removed `.clone()` on Copy `DateTime`,
dropped no-op i32->i32 casts, `assert_eq!(x, bool)` -> `assert!`,
`if let Ok` over iterators -> `.flatten()`, `map_or(true, ..)` ->
`is_none_or`, redundant closures, a collapsible `if`, and moved the
stray `impl From<Instance>` above the ec2 test module. Wide AWS
parameter sets that legitimately exceed the arg-count threshold carry
`#[allow(clippy::too_many_arguments)]`, and the deprecated Cognito
`UserPoolType::status` read carries `#[allow(deprecated)]`.

+++ src/schema/direct_connect/types.rs
@@ -52,7 +52,7 @@ impl From<VirtualInterface> for DxVirtualInterface {
             virtual_interface_state: v.virtual_interface_state().map(|s| s.as_str().to_string()),
             connection_id: v.connection_id().map(|s| s.to_string()),
             vlan: if v.vlan() == 0 { None } else { Some(v.vlan()) },
-            asn: Some(v.asn() as i32),
+            asn: Some(v.asn()),
             amazon_address: v.amazon_address().map(|s| s.to_string()),
             customer_address: v.customer_address().map(|s| s.to_string()),
         }
